# [ready] perf: faster lazyop init