@tinyb0t 2023-08-25
Changes made in `tinygrad/`:
```
------------------------------------------------------------
files                             insertions       deletions
------------------------------------------------------------
tinygrad/ops.py                            2               1
------------------------------------------------------------
lines added in the tinygrad folder: 1
```