@roelofvandijk 2023-08-26
@wozeparrot Updated.

```
original

Total time: 1.01442 s
File: /home/rvd/src/roelofvandijk/tinygrad/tinygrad/ops.py
Function: __init__ at line 28

Line #      Hits         Time  Per Hit   % Time  Line Contents
==============================================================
    28                                             @profile
    29                                             def __init__(self, op: Op, src: Tuple[Union[LazyOp, LazyBuffer], ...], arg: Any = None):
    30    377351     100565.1      0.3      9.9      self.op = op
    31    377351      74519.3      0.2      7.3      self.src = src
    32    377351      56566.0      0.1      5.6      self.arg = arg
    33    377351      44155.7      0.1      4.4      try:
    34    306065     698992.6      2.3     68.9        self.buffers = tuple([y for x in src for y in x.buffers])
    35     71286      15703.6      0.2      1.5      except AttributeError:
    36                                                 # NOTE: the linearizer's key function maps the buffers to ints, and LOCAL_BUFFER is used. we don't care about buffers in these cases
    37     71286      23921.0      0.3      2.4        pass




only buffers line change
Total time: 0.750368 s
File: /home/rvd/src/roelofvandijk/tinygrad/tinygrad/ops.py
Function: __init__ at line 29

Line #      Hits         Time  Per Hit   % Time  Line Contents
==============================================================
    29                                             @profile
    30                                             def __init__(self, op: Op, src: Tuple[Union[LazyOp, LazyBuffer], ...], arg: Any = None):
    31    377351      78436.6      0.2     10.5      self.op = op
    32    377351      68671.9      0.2      9.2      self.src = src
    33    377351      58733.7      0.2      7.8      self.arg = arg
    34    377351      44611.2      0.1      5.9      try:
    35    377351      62295.1      0.2      8.3        self.buffers = ()
    36    803341     421537.2      0.5     56.2        for x in src: self.buffers += x.buffers
    37     32586       7652.5      0.2      1.0      except AttributeError:
    38                                                 # NOTE: the linearizer's key function maps the buffers to ints, and LOCAL_BUFFER is used. we don't care about buffers in these cases
    39     32586       8429.5      0.3      1.1        pass



grouped init (this should become a dataclass)
Total time: 0.692713 s
File: /home/rvd/src/roelofvandijk/tinygrad/tinygrad/ops.py
Function: __init__ at line 29

Line #      Hits         Time  Per Hit   % Time  Line Contents
==============================================================
    29                                             @profile
    30                                             def __init__(self, op: Op, src: Tuple[Union[LazyOp, LazyBuffer], ...], arg: Any = None):
    31    377351     195652.9      0.5     28.2      self.op, self.src, self.arg, self.buffers = op, src, arg, ()
    32    377351      61017.4      0.2      8.8      try:
    33    803341     418867.6      0.5     60.5        for x in src: self.buffers += x.buffers
    34     32586       8202.6      0.3      1.2      except AttributeError:
    35                                                 # NOTE: the linearizer's key function maps the buffers to ints, and LOCAL_BUFFER is used. we don't care about buffers in these cases
    36     32586       8972.9      0.3      1.3        pass

```